backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
faiss-cpu
pypdf
sentence-transformers
reportlab
python-dotenv
pydantic
httpx
